refactor: 체크 스타일에 맞게 수정
1. Bean 등록을 AppConfig로 이동
2. 띄어쓰기
3. Json의 대문자 변수
