+++ src/main/java/kr/co/fastcampus/fastcatch/domain/cartitem/entity/CartItem.java
@@ -1,4 +1,4 @@
-package kr.co.fastcampus.fastcatch.domain.order_item.entity;
+package kr.co.fastcampus.fastcatch.domain.cartitem.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.Column;
@@ -8,10 +8,9 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import java.time.LocalDate;
 import kr.co.fastcampus.fastcatch.common.baseentity.BaseEntity;
-import kr.co.fastcampus.fastcatch.domain.orders.entity.Orders;
+import kr.co.fastcampus.fastcatch.domain.cart.entity.Cart;
 import kr.co.fastcampus.fastcatch.domain.room.entity.Room;
 import lombok.Builder;
 import lombok.Getter;
@@ -20,12 +19,12 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @Entity
-public class OrderItem extends BaseEntity {
+public class CartItem extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @JsonIgnore
-    private Long orderItemId;
+    private Long cartItemId;
 
     @Column(nullable = false)
     private LocalDate startDate;
@@ -40,22 +39,22 @@ public class OrderItem extends BaseEntity {
     private Integer price;
 
     @ManyToOne
-    @JoinColumn(name = "order_id")
+    @JoinColumn(name = "cart_id")
     @JsonIgnore
-    private Orders orders;
+    private Cart cart;
 
-    @OneToOne
+    @ManyToOne
     @JoinColumn(name = "room_id")
     @JsonIgnore
     private Room room;
 
     @Builder
-    public OrderItem(
-        Long orderItemId, Orders orders, Room room, LocalDate startDate,
+    public CartItem(
+        Long cartItemId, Cart cart, Room room, LocalDate startDate,
         LocalDate endDate, Integer headCount, Integer price
     ) {
-        this.orderItemId = orderItemId;
-        this.orders = orders;
+        this.cartItemId = cartItemId;
+        this.cart = cart;
         this.room = room;
         this.startDate = startDate;
         this.endDate = endDate;
