+++ src/main/java/kr/co/fastcampus/fastcatch/domain/orderitem/entity/OrderItem.java
@@ -1,4 +1,4 @@
-package kr.co.fastcampus.fastcatch.domain.cart_item.entity;
+package kr.co.fastcampus.fastcatch.domain.orderitem.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.Column;
@@ -8,9 +8,10 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import java.time.LocalDate;
 import kr.co.fastcampus.fastcatch.common.baseentity.BaseEntity;
-import kr.co.fastcampus.fastcatch.domain.cart.entity.Cart;
+import kr.co.fastcampus.fastcatch.domain.orders.entity.Orders;
 import kr.co.fastcampus.fastcatch.domain.room.entity.Room;
 import lombok.Builder;
 import lombok.Getter;
@@ -19,12 +20,12 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @Entity
-public class CartItem extends BaseEntity {
+public class OrderItem extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @JsonIgnore
-    private Long cartItemId;
+    private Long orderItemId;
 
     @Column(nullable = false)
     private LocalDate startDate;
@@ -39,22 +40,22 @@ public class CartItem extends BaseEntity {
     private Integer price;
 
     @ManyToOne
-    @JoinColumn(name = "cart_id")
+    @JoinColumn(name = "order_id")
     @JsonIgnore
-    private Cart cart;
+    private Orders orders;
 
-    @ManyToOne
+    @OneToOne
     @JoinColumn(name = "room_id")
     @JsonIgnore
     private Room room;
 
     @Builder
-    public CartItem(
-        Long cartItemId, Cart cart, Room room, LocalDate startDate,
+    public OrderItem(
+        Long orderItemId, Orders orders, Room room, LocalDate startDate,
         LocalDate endDate, Integer headCount, Integer price
     ) {
-        this.cartItemId = cartItemId;
-        this.cart = cart;
+        this.orderItemId = orderItemId;
+        this.orders = orders;
         this.room = room;
         this.startDate = startDate;
         this.endDate = endDate;
