+++ src/main/java/kr/co/fastcampus/fastcatch/domain/orders/entity/Orders.java
@@ -17,7 +17,7 @@ import java.util.ArrayList;
 import java.util.List;
 import kr.co.fastcampus.fastcatch.common.baseentity.BaseEntity;
 import kr.co.fastcampus.fastcatch.domain.member.entity.Member;
-import kr.co.fastcampus.fastcatch.domain.order_item.entity.OrderItem;
+import kr.co.fastcampus.fastcatch.domain.orderitem.entity.OrderItem;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
