+++ src/main/resources/Accommodation.json
@@ -1,5 +1,5 @@
 {
- "DATA": [
+ "data": [
   {
    "id": 1,
    "name": "토요코 인 서면",
